Revert "fix(docs): Revert include aztec rpc interface typedoc output in docs (#2255)"
This reverts commit f85243298ef2a5c01764e592c6f6ea50d835bf07.
