+++ .circleci/config.yml
@@ -1315,8 +1315,6 @@ workflows:
     when:
       equal: [system, << pipeline.parameters.workflow >>]
     jobs:
-      - build-docs: *defaults
-
       # Barretenberg
       - barretenberg-x86_64-linux-gcc: *defaults
       - barretenberg-x86_64-linux-clang: *defaults
@@ -1409,7 +1407,8 @@ workflows:
             - l1-contracts
             - noir-contracts-build
           <<: *defaults
-
+      
+      - build-docs: *yarn_project
       - aztec-js: *yarn_project
       - end-to-end: *yarn_project
       - ethereum: *yarn_project
