+++ docs/Dockerfile
@@ -1,10 +1,14 @@
-FROM node:18-alpine
-RUN apk update
+FROM 278380418400.dkr.ecr.eu-west-2.amazonaws.com/yarn-project-base AS builder
 
 WORKDIR /usr/src
 
 COPY . .
 
+WORKDIR /usr/src/yarn-project
+
+RUN yarn build
+
 WORKDIR /usr/src/docs
 
 RUN yarn && yarn build
+
