+++ yarn-project/types/src/interfaces/index.ts
@@ -4,3 +4,6 @@ export * from './data_commitment_provider.js';
 export * from './l1_l2_message_provider.js';
 export * from './aztec-node.js';
 export * from './aztec_rpc.js';
+export * from './deployed-contract.js';
+export * from './node-info.js';
+export * from './sync-status.js';
